requirements.txt
absl-py==2.4.0
altair==6.0.0
annotated-doc==0.0.4
anyio==4.13.0
astunparse==1.6.3
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
filelock==3.28.0
Flask==3.1.3
flatbuffers==25.12.19
fsspec==2026.3.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.80.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.14.0
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.19.0
numpy==2.4.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.1
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
pyarrow==23.0.1
pydeck==0.9.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
wheel==0.46.3
wrapt==2.1.2
